+++ lib/src/analyzers/lint_analyzer/rules/rules_list/avoid_redundant_async_on_load/avoid_redundant_async_on_load_rule.dart
@@ -2,25 +2,26 @@
 
 import 'package:analyzer/dart/ast/ast.dart';
 import 'package:analyzer/dart/ast/visitor.dart';
-import 'package:analyzer/dart/element/element.dart';
+import 'package:analyzer/dart/element/type.dart';
+import 'package:collection/collection.dart';
 
+import '../../../../../utils/flame_type_utils.dart';
 import '../../../../../utils/node_utils.dart';
 import '../../../lint_utils.dart';
 import '../../../models/internal_resolved_unit_result.dart';
 import '../../../models/issue.dart';
 import '../../../models/severity.dart';
-import '../../models/common_rule.dart';
+import '../../models/flame_rule.dart';
 import '../../rule_utils.dart';
 
 part 'visitor.dart';
 
-class AvoidGlobalStateRule extends CommonRule {
-  static const String ruleId = 'avoid-global-state';
+class AvoidRedundantAsyncOnLoadRule extends FlameRule {
+  static const String ruleId = 'avoid-redundant-async-on-load';
 
-  static const _warning =
-      'Avoid using global variable without const or final keywords.';
+  static const _warningMessage = "Avoid unnecessary async 'onLoad'.";
 
-  AvoidGlobalStateRule([Map<String, Object> config = const {}])
+  AvoidRedundantAsyncOnLoadRule([Map<String, Object> config = const {}])
       : super(
           id: ruleId,
           severity: readSeverity(config, Severity.warning),
@@ -38,7 +39,7 @@ class AvoidGlobalStateRule extends CommonRule {
         .map((declaration) => createIssue(
               rule: this,
               location: nodeLocation(node: declaration, source: source),
-              message: _warning,
+              message: _warningMessage,
             ))
         .toList(growable: false);
   }